Refactor interfaces, in particular separate Solution.
