+++ CMakeLists_files.cmake
@@ -28,6 +28,7 @@ list (APPEND MAIN_SOURCE_FILES
         opm/flowdiagnostics/CellSetValues.cpp
         opm/flowdiagnostics/ConnectionValues.cpp
         opm/flowdiagnostics/ConnectivityGraph.cpp
+        opm/flowdiagnostics/Solution.cpp
         opm/flowdiagnostics/Toolbox.cpp
         opm/flowdiagnostics/TracerTofSolver.cpp
         opm/utility/graph/tarjan.c
@@ -50,6 +51,7 @@ list (APPEND PUBLIC_HEADER_FILES
         opm/flowdiagnostics/CellSetValues.hpp
         opm/flowdiagnostics/ConnectionValues.hpp
         opm/flowdiagnostics/ConnectivityGraph.hpp
+        opm/flowdiagnostics/Solution.hpp
         opm/flowdiagnostics/Toolbox.hpp
         opm/flowdiagnostics/TracerTofSolver.hpp
         opm/utility/graph/AssembledConnections.hpp
